[NFC] Deduplicate comment in PromoteMemoryToRegister.cpp
This has been duplicated since before
eee7f4fe6ff6b89d21aebb2ff0747f68b9c8b85b, but that commit has it
appearing twice in the space of 10 lines of the same function body. It
could also be hoisted up to the point just after where the last
special-case is considered, but I want to keep the intent of the
original authors.

Committed as obvious without a review.