Re-land [Debug][CodeView] Emit fully qualified names for globals
This reverts commit 676fb0f56516c3c2ff84ea87ca94fa594f3bad51.

Fixed an issue with pointers to members based on typedefs. In this case,
LLVM would emit a second UDT. I fixed it by not passing the class type
to getTypeIndex when the base type is not a function type. lowerType
only uses the class type for direct function types. This suggests if we
have a PMF with a function typedef, there may be an issue, but that can
be solved separately.